refactor(hw_41): optimization files
