+++ react/my-app/src/App.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import logo from './logo.svg';
 import { useState } from 'react';
 import './App.css';
-import { Wrapper } from './hw_37/wrapper';
-import { ChangeTheme, ThemeContext } from './hw_41/themeContext';
+import { ChangeTheme, ThemeContext } from './tools/themeContext';
+import { Wrapper } from './components/wrapper';
 
 function App() {
