+++ react/my-app/src/components/PostsWrapper.tsx
