+++ react/my-app/src/components/Registration.tsx
