+++ react/my-app/src/styles/reg.css
