+++ react/my-app/src/tools/getPosts.ts
